ClimascalienteS11.3/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
schedule==1.2.0